requirements.txt
IPython
requests
pandas
mysql-connector-python